requirements.txt
streamlit
torch
transformers
diffusers
accelerate
sentence-transformers
faiss-cpu
ftfy
bitsandbytes

-e .